+++ logger.go
@@ -58,18 +58,28 @@ func New() *Logger {
 	}
 }
 
+// Fields is just a custom type of the map type.
+// One or more values of `Fields` type can be passed
+// on all Log methods except `Print/Printf/Println` to set the `Log.Fields` field,
+// which can be accessed through a custom LogHandler.
+type Fields map[string]interface{}
+
 // acquireLog returns a new log fom the pool.
-func (l *Logger) acquireLog(level Level, msg string, withPrintln bool) *Log {
+func (l *Logger) acquireLog(level Level, msg string, withPrintln bool, fields Fields) *Log {
 	log, ok := l.logs.Get().(*Log)
 	if !ok {
 		log = &Log{
 			Logger: l,
 		}
 	}
+
 	log.NewLine = withPrintln
 	log.Time = time.Now()
+	log.Timestamp = log.Time.Unix()
 	log.Level = level
 	log.Message = msg
+	log.Fields = fields
+	log.Stacktrace = log.Stacktrace[:0]
 	return log
 }
 
@@ -198,12 +208,15 @@ func (l *Logger) SetLevel(levelName string) *Logger {
 	return l
 }
 
-func (l *Logger) print(level Level, msg string, newLine bool) {
+func (l *Logger) print(level Level, msg string, newLine bool, fields Fields) {
 	if l.Level >= level {
 		// newLine passed here in order for handler to know
 		// if this message derives from Println and Leveled functions
 		// or by simply, Print.
-		log := l.acquireLog(level, msg, newLine)
+		log := l.acquireLog(level, msg, newLine, fields)
+		if level == DebugLevel {
+			log.Stacktrace = GetStacktrace()
+		}
 		// if not handled by one of the handler
 		// then print it as usual.
 		if !l.handled(log) {
@@ -224,33 +237,64 @@ func (l *Logger) print(level Level, msg string, newLine bool) {
 
 // Print prints a log message without levels and colors.
 func (l *Logger) Print(v ...interface{}) {
-	l.print(DisableLevel, fmt.Sprint(v...), l.NewLine)
+	l.print(DisableLevel, fmt.Sprint(v...), l.NewLine, nil)
 }
 
 // Printf formats according to a format specifier and writes to `Printer#Output` without levels and colors.
 func (l *Logger) Printf(format string, args ...interface{}) {
-	l.print(DisableLevel, fmt.Sprintf(format, args...), l.NewLine)
+	l.print(DisableLevel, fmt.Sprintf(format, args...), l.NewLine, nil)
 }
 
 // Println prints a log message without levels and colors.
 // It adds a new line at the end, it overrides the `NewLine` option.
 func (l *Logger) Println(v ...interface{}) {
-	l.print(DisableLevel, fmt.Sprint(v...), true)
+	l.print(DisableLevel, fmt.Sprint(v...), true, nil)
+}
+
+func splitArgsFields(v []interface{}) ([]interface{}, Fields) {
+	var (
+		args   = v[:0]
+		fields Fields
+	)
+
+	for _, val := range v {
+		f, ok := val.(Fields)
+		if ok {
+			if len(fields) > 0 { // upsert new values.
+				for k, v := range f {
+					fields[k] = v
+				}
+			} else {
+				fields = f // set fields.
+			}
+			continue
+		}
+
+		args = append(args, val) // use it as fmt argument.
+	}
+
+	return args, fields
 }
 
 // Log prints a leveled log message to the output.
 // This method can be used to use custom log levels if needed.
 // It adds a new line in the end.
 func (l *Logger) Log(level Level, v ...interface{}) {
-	l.print(level, fmt.Sprint(v...), l.NewLine)
+	if l.Level >= level {
+		args, fields := splitArgsFields(v)
+		l.print(level, fmt.Sprint(args...), l.NewLine, fields)
+	}
 }
 
 // Logf prints a leveled log message to the output.
 // This method can be used to use custom log levels if needed.
 // It adds a new line in the end.
 func (l *Logger) Logf(level Level, format string, args ...interface{}) {
-	msg := fmt.Sprintf(format, args...)
-	l.Log(level, msg)
+	if l.Level >= level {
+		args, fields := splitArgsFields(args)
+		msg := fmt.Sprintf(format, args...)
+		l.print(level, msg, l.NewLine, fields)
+	}
 }
 
 // Fatal `os.Exit(1)` exit no matter the level of the logger.
@@ -264,8 +308,7 @@ func (l *Logger) Fatal(v ...interface{}) {
 // If the logger's level is fatal, error, warn, info or debug
 // then it will print the log message too.
 func (l *Logger) Fatalf(format string, args ...interface{}) {
-	msg := fmt.Sprintf(format, args...)
-	l.Fatal(msg)
+	l.Logf(FatalLevel, format, args...)
 }
 
 // Error will print only when logger's Level is error, warn, info or debug.
@@ -275,8 +318,7 @@ func (l *Logger) Error(v ...interface{}) {
 
 // Errorf will print only when logger's Level is error, warn, info or debug.
 func (l *Logger) Errorf(format string, args ...interface{}) {
-	msg := fmt.Sprintf(format, args...)
-	l.Error(msg)
+	l.Logf(ErrorLevel, format, args...)
 }
 
 // Warn will print when logger's Level is warn, info or debug.
@@ -286,8 +328,7 @@ func (l *Logger) Warn(v ...interface{}) {
 
 // Warnf will print when logger's Level is warn, info or debug.
 func (l *Logger) Warnf(format string, args ...interface{}) {
-	msg := fmt.Sprintf(format, args...)
-	l.Warn(msg)
+	l.Logf(WarnLevel, format, args...)
 }
 
 // Warningf exactly the same as `Warnf`.
@@ -304,8 +345,7 @@ func (l *Logger) Info(v ...interface{}) {
 
 // Infof will print when logger's Level is info or debug.
 func (l *Logger) Infof(format string, args ...interface{}) {
-	msg := fmt.Sprintf(format, args...)
-	l.Info(msg)
+	l.Logf(InfoLevel, format, args...)
 }
 
 // Debug will print when logger's Level is debug.
@@ -315,13 +355,7 @@ func (l *Logger) Debug(v ...interface{}) {
 
 // Debugf will print when logger's Level is debug.
 func (l *Logger) Debugf(format string, args ...interface{}) {
-	// On debug mode don't even try to fmt.Sprintf if it's not required,
-	// this can be used to allow `Debugf` to be called without even the `fmt.Sprintf`'s
-	// performance cost if the logger doesn't allow debug logging.
-	if l.Level >= DebugLevel {
-		msg := fmt.Sprintf(format, args...)
-		l.Debug(msg)
-	}
+	l.Logf(DebugLevel, format, args...)
 }
 
 // Install receives  an external logger